+++ packages/@expo/metro-runtime/src/effects.native.ts
@@ -1,9 +1,0 @@
-// Only during development.
-if (process.env.NODE_ENV !== 'production') {
-  if (
-    // Disable for SSR
-    typeof globalThis.expo !== 'undefined'
-  ) {
-    require('./messageSocket');
-  }
-}