+++ packages/expo/src/Expo.fx.web.tsx
@@ -1,4 +1,16 @@
 import './winter';
+import './async-require';
+
+// Only during development.
+if (
+  __DEV__ &&
+  // Disable for SSR
+  typeof window !== 'undefined'
+) {
+  require('./async-require/setupFastRefresh');
+  require('./async-require/setupHMR');
+  require('./async-require/messageSocket');
+}
 
 // When users dangerously import a file inside of react-native, it breaks the web alias.
 // This is one of the most common, and cryptic web errors that users encounter.