+++ packages/expo/src/async-require/loadBundle.ts
@@ -19,7 +19,7 @@ export async function loadBundleAsync(bundlePath: string): Promise<void> {
     return fetchThenEvalAsync(requestUrl);
   } else {
     return fetchThenEvalAsync(requestUrl).then(() => {
-      const HMRClient = require('../HMRClient').default as typeof import('../HMRClient').default;
+      const HMRClient = require('./hmr').default as typeof import('./hmr').default;
       HMRClient.registerBundle(requestUrl);
     });
   }